视图大小调整

+++ ShiShiCai/MainWindow.xaml.cs
@@ -180,7 +180,7 @@ namespace ShiShiCai
 
                     item.FullValue = dr["C002"].ToString();
                     item.LargeValue = dr["C007"].ToString() == "1";
-                    item.DoubleValue = dr["C008"].ToString() == "1";
+                    item.SingleValue = dr["C008"].ToString() == "1";
                     item.SumValue = Convert.ToInt32(dr["C009"]);
                     item.RepeatValue = dr["C100"].ToString() == "1";
                     item.IntervalValue = dr["C101"].ToString() == "1";
@@ -233,7 +233,7 @@ namespace ShiShiCai
 
                     item.FullValue = dr["C002"].ToString();
                     item.LargeValue = dr["C007"].ToString() == "2";
-                    item.DoubleValue = dr["C008"].ToString() == "2";
+                    item.SingleValue = dr["C008"].ToString() == "2";
                     item.SumValue = Convert.ToInt32(dr["C009"]);
                     item.RepeatValue = dr["C100"].ToString() == "2";
                     item.IntervalValue = dr["C101"].ToString() == "2";
